Modulo 4 - Desarrollo Web/MP3/requirements.txt
django==2.2.20
djangorestframework
drf-yasg
requests
